refactor(tests): Move vis tests from API to vis
